merge (dev => main): update main branch

+++ GameEngine/src/Exodia/ImGui/ImGuiLayer.cpp
@@ -13,6 +13,7 @@
     #include <backends/imgui_impl_glfw.h>
     #include <backends/imgui_impl_opengl3.h>
     #include <imgui_internal.h>
+    #include <ImGuizmo.h>
 
 // Exodia Core
 #include "Core/Application/Application.hpp"
@@ -95,7 +96,7 @@ namespace Exodia {
         if (_BlockEvents) {
             ImGuiIO &io = ImGui::GetIO();
 
-            event.Handled |= event.IsInCategory(EventCategory::EventCategoryMouse) & io.WantCaptureMouse;
+            event.Handled |= event.IsInCategory(EventCategory::EventCategoryMouse)    & io.WantCaptureMouse;
             event.Handled |= event.IsInCategory(EventCategory::EventCategoryKeyboard) & io.WantCaptureKeyboard;
         }
     }
@@ -118,7 +119,7 @@ namespace Exodia {
         ImGuiIO &io = ImGui::GetIO();
         Application &app = Application::Get();
 
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
         // Rendering
         ImGui::Render();
